examples/FoR/BlocksWorld/LLMs-Planning/llm_planning_analysis/requirements.txt
numpy==1.19.5
openai
PyYAML==6.0
tarski==0.7.0
transformers==4.22.2
pddl==0.2.0